Move winds.py code to Sensor in inputs. Delete winds.py
