+++ quicfire_tools/inputs.py
@@ -2061,14 +2061,17 @@ class QU_metparams(InputFile):
         )
 
 
-class Sensor1(InputFile):
+class Sensor(InputFile):
     """
-    Class representing the sensor1.inp input file.
+    Class representing a sensor*.inp input file.
     This file contains information on winds, and serves as the
-    primary source for wind speed(s) and direction(s)
+    primary source for wind speed(s) and direction(s). 
+    Multiple sensor*.inp files may be created.
 
     Attributes
     ----------
+    sensor_number : PositiveInt
+        Which numbered sensor does this class represent #TODO better description
     time_now : PositiveInt
         Begining of time step in Unix Epoch time (integer seconds since
         1970/1/1 00:00:00). Must match time at beginning of fire
@@ -2081,30 +2084,50 @@ class Sensor1(InputFile):
         Wind direction (degrees). Use 0° for North
     """
 
-    name: str = "sensor1"
+    sensor_number: PositiveInt = 1
     _extension: str = ".inp"
     time_now: PositiveInt
+    wind_times: Union[NonNegativeFloat,list(NonNegativeFloat)]
+    wind_speeds: Union[PositiveFloat, list(PositiveFloat)]
+    wind_directions: Union[PositiveFloat, list(PositiveFloat)]
     sensor_height: PositiveFloat = 6.1
-    wind_speed: PositiveFloat
-    wind_direction: NonNegativeInt = Field(lt=360)
+    x_location: PositiveInt
+    y_location: PositiveInt
+
+    @computed_field
+    @property
+    def name(self) -> str:
+        return "sensor" + self.sensor_number
+    
+    def _validate_wind_lists(self):
+        if isinstance(self.wind_times,float): self.wind_times = [self.wind_times]
+        if isinstance(self.wind_speeds, float): self.wind_speeds = [self.wind_speeds]
+        if isinstance(self.wind_directions, float): self.wind_directions = [self.wind_directions]
+        if len(self.wind_times) != len(self.wind_speeds) != len(self.wind_directions): #TODO figure out how to do this right
+            raise ValueError(f"WindSensor: lists of wind times, speeds, and directions must be the same length.\n",
+                             f"len(wind_times) = {len(self.wind_times)}\n",
+                             f"len(wind_speeds) = {len(self.wind_speeds)}\n",
+                             f"len(win_directions) = {len(self.wind_directions)}")
 
     @computed_field
     @property
     def _wind_lines(self) -> str:
-        """
-        This is meant to support wind shifts in the future.
-        This computed field could be altered to reproduce the lines below
-        for a series of times, speeds, and directions.
-        """
-        return (
-            f"{self.time_now} !Begining of time step in Unix Epoch time\n"
-            f"1 !site boundary layer flag (1 = log, 2 = exp, 3 = urban canopy, 4 = discrete data points)\n"
-            f"0.1 !site zo\n"
-            f"0. ! 1/L (default = 0)\n"
-            f"!Height (m), Speed (m/s), Direction (deg relative to true N)\n"
-            f"{self.sensor_height} {self.wind_speed} {self.wind_direction}"
-        )
+        self._validate_wind_lists()
+        location_lines = (f"{self.x_location} !X coordinate (meters)\n",
+                          f"{self.y_location} !Y coordinate (meters)\n")
+        windshifts = []
+        for i in len(self.wind_times):
+            shift = (f"\n{self.time_now + self.wind_times[i]} !Begining of time step in Unix Epoch time (integer seconds since 1970/1/1 00:00:00)\n",
+                     f"1 !site boundary layer flag (1 = log, 2 = exp, 3 = urban canopy, 4 = discrete data points)\n",
+                     f"0.1 !site zo\n",
+                     f"0. ! 1/L (default = 0)\n"
+                     f"!Height (m),Speed	(m/s), Direction (deg relative to true N)\n"
+                     f"{self.sensor_height} {self.wind_speeds[i]} {self.wind_directions[i]}")
+            windshifts.append(shift)
+        
+        return location_lines + "".join(windshifts)
 
+    # TODO Update from_file
     @classmethod
     def from_file(cls, directory: str | Path):
         if isinstance(directory, str):
@@ -2118,3 +2141,18 @@ class Sensor1(InputFile):
             wind_speed=float(lines[11].split(" ")[1]),
             wind_direction=int(lines[11].split(" ")[2]),
         )
+
+    @classmethod
+    def from_csv(cls,
+                 filename: Union[str,Path]):
+        """
+        Create windshifts from a .csv file.
+        
+        """
+        if isinstance(filename, str):
+            filename = Path(filename)
+       
+        with open(filename,"r") as f:
+            matrix = list(csv.reader(f))
+        
+        # TODO: add function for calculating windspeed/direction from U,V winds. Put in quicfire_tools.utils
